Yoga-Pose-Detection/asana-bridge/backend/requirements.txt
# Backend Requirements - Asana-Bridge with MoveNet
# Python 3.10+

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ML & Computer Vision
tensorflow>=2.15.0
tensorflow-hub>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# Utilities
pydantic>=2.5.0
